修改网页分析  实体类属性的数据类型   并添加无参构造函数（mybatis返回resultType时  要求对应实体类有无参构造函数）

+++ src/main/java/com/trs/gov/kpi/entity/PageDepth.java
@@ -18,7 +18,7 @@ public class PageDepth {
 
     private Integer depth;
 
-    private Integer space;
+    private Long space;
 
     private Date checkTime;
 }
